Implementa Copy service, serviceImpl y Controller

+++ src/main/java/com/videostore/modules/copy/repository/CopyRepository.java
@@ -3,5 +3,5 @@ package com.videostore.modules.copy.repository;
 import com.videostore.modules.copy.model.Copy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CopyRepository extends JpaRepository<Copy, Long> {
+public interface CopyRepository extends JpaRepository<Copy, Integer> {
 }
